Для всех численных ID заменил int на int64.

+++ services/auth/internal/usecase/session.go
@@ -14,7 +14,7 @@ import (
 //go:generate mockgen -destination=mocks/session_mock.go -package=mocks github.com/go-park-mail-ru/2026_1_NaNcats/internal/usecase/auth SessionUseCase
 type SessionUseCase interface {
 	// бизнес-логика создания сессии для пользователя, вовзращает sessionID
-	Create(ctx context.Context, userID int, userAgent string) (domain.Session, error)
+	Create(ctx context.Context, userID int64, userAgent string) (domain.Session, error)
 	// проверяет, существует и не истек ли sessionID, возвращает айди юзера при успехе
 	Check(ctx context.Context, id uuid.UUID) (domain.Session, error)
 	// бизнес-логика для удаления сессии, просто вызывает удаление из repository.session
@@ -36,7 +36,7 @@ func NewSessionUseCase(sr repository.SessionRepository, ttl time.Duration) Sessi
 	}
 }
 
-func (u *sessionUseCase) Create(ctx context.Context, userID int, userAgent string) (domain.Session, error) {
+func (u *sessionUseCase) Create(ctx context.Context, userID int64, userAgent string) (domain.Session, error) {
 	// бизнес-логика создания сессии
 	// возвращает sessionID созданной сессии и момент времени, когда истекает
